# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NOVA Ride API",
    "version": "1.0.0",
    "description": "جرد مسارات مُولّد آليًا من المتحكّمات. البادئة /api؛ كل المسارات متاحة أيضًا تحت /api/v1. لا يتضمّن هذا الجرد مخطّطات الطلب/الاستجابة (تُثرى لاحقًا)."
  },
  "servers": [
    {
      "url": "{baseUrl}",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:4000"
        }
      }
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/driver/trips/:id/arriving": {
      "post": {
        "tags": [
          "trips"
        ],
        "operationId": "trips_arriving",
        "summary": "POST /api/driver/trips/:id/arriving",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/driver/trips/:id/cancel": {
      "post": {
        "tags": [
          "trips"
        ],
        "operationId": "trips_cancel",
        "summary": "POST /api/driver/trips/:id/cancel",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/driver/trips/:id/complete": {
      "post": {
        "tags": [
          "trips"
        ],
        "operationId": "trips_complete",
        "summary": "POST /api/driver/trips/:id/complete",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/driver/trips/:id/start": {
      "post": {
        "tags": [
          "trips"
        ],
        "operationId": "trips_start",
        "summary": "POST /api/driver/trips/:id/start",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/driver/trips/:id/track": {
      "get": {
        "tags": [
          "trips"
        ],
        "operationId": "trips_track",
        "summary": "GET /api/driver/trips/:id/track",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/setting-change-requests": {
      "get": {
        "tags": [
          "settings"
        ],
        "operationId": "settings_findAll",
        "summary": "GET /api/setting-change-requests",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      }
    },
    "/api/setting-change-requests/:id/approve": {
      "post": {
        "tags": [
          "settings"
        ],
        "operationId": "settings_approve",
        "summary": "POST /api/setting-change-requests/:id/approve",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      }
    },
    "/api/setting-change-requests/:id/reject": {
      "post": {
        "tags": [
          "settings"
        ],
        "operationId": "settings_reject",
        "summary": "POST /api/setting-change-requests/:id/reject",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      }
    },
...
    "/api/subscriptions": {
      "get": {
        "tags": [
          "subscriptions"
        ],
        "operationId": "subscriptions_findAll",
        "summary": "GET /api/subscriptions",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: subscriptions.manage"
      }
    },
    "/api/subscriptions/cancel": {
      "post": {
        "tags": [
          "subscriptions"
        ],
        "operationId": "subscriptions_cancel",
        "summary": "POST /api/subscriptions/cancel",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/subscriptions/me": {
      "get": {
        "tags": [
          "subscriptions"
        ],
        "operationId": "subscriptions_me",
        "summary": "GET /api/subscriptions/me",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/subscriptions/plans": {
      "get": {
        "tags": [
          "subscriptions"
        ],
        "operationId": "subscriptions_plans",
        "summary": "GET /api/subscriptions/plans",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "subscriptions"
        ],
        "operationId": "subscriptions_createPlan",
        "summary": "POST /api/subscriptions/plans",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: subscriptions.manage"
      }
    },
    "/api/subscriptions/plans/:id": {
      "delete": {
        "tags": [
          "subscriptions"
        ],
        "operationId": "subscriptions_deactivatePlan",
        "summary": "DELETE /api/subscriptions/plans/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: subscriptions.manage"
      },
      "patch": {
        "tags": [
          "subscriptions"
        ],
        "operationId": "subscriptions_updatePlan",
        "summary": "PATCH /api/subscriptions/plans/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: subscriptions.manage"
      }
    },
    "/api/subscriptions/plans/all": {
      "get": {
        "tags": [
          "subscriptions"
        ],
        "operationId": "subscriptions_allPlans",
        "summary": "GET /api/subscriptions/plans/all",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: subscriptions.manage"
      }
    },
    "/api/subscriptions/subscribe": {
      "post": {
        "tags": [
          "subscriptions"
        ],
        "operationId": "subscriptions_subscribe",
        "summary": "POST /api/subscriptions/subscribe",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/support/tickets": {
      "get": {
        "tags": [
          "support"
        ],
        "operationId": "support_all",
        "summary": "GET /api/support/tickets",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: support.manage"
      },
      "post": {
        "tags": [
          "support"
        ],
        "operationId": "support_create",
        "summary": "POST /api/support/tickets",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/support/tickets/:id/assign": {
      "patch": {
        "tags": [
          "support"
        ],
        "operationId": "support_assign",
        "summary": "PATCH /api/support/tickets/:id/assign",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/support/tickets/:id/first-response": {
      "post": {
        "tags": [
          "support"
        ],
        "operationId": "support_firstResponse",
        "summary": "POST /api/support/tickets/:id/first-response",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/support/tickets/:id/priority": {
      "patch": {
        "tags": [
          "support"
        ],
        "operationId": "support_setPriority",
        "summary": "PATCH /api/support/tickets/:id/priority",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/support/tickets/:id/resolve": {
      "patch": {
        "tags": [
          "support"
        ],
        "operationId": "support_resolve",
        "summary": "PATCH /api/support/tickets/:id/resolve",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/support/tickets/breaching": {
      "get": {
        "tags": [
          "support"
        ],
        "operationId": "support_breaching",
        "summary": "GET /api/support/tickets/breaching",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}